Added the following applications: Salary 9. Write an application that calculates and displays the weekly salary for an employee. The main() method prompts the user for an hourly pay rate, regular hours, and overtime hours. Create a separate method to calculate overtime pay, which is regular hours times the pay rate plus overtime hours times 1.5 times the pay rate; return the result to the main() method to be displayed. Save the program as Salary.java.
Interest
10. Write an application that calculates and displays the amount of money a user would have if his or her money
could be invested at 5 percent interest for one year. Create a method that prompts the user for the starting
value of the investment and returns it to the calling program. Call a separate method to do the calculation, and
return the result to be displayed. Save the program as Interest.java.

Sandwich
11. a. Create a class named Sandwich. Data fields include a String for the main ingredient (such as “tuna”), a
String for bread type (such as “wheat”), and a double for price (such as 4.99). Include methods to get and set
values for each of these fields. Save the class as Sandwich.java.

TestSandwicxh
b. Create an application named TestSandwich that instantiates one Sandwich object and demonstrates the use
of the set and get methods. Save this application as TestSandwich.java.

Student
12. a. Create a class named Student. A Student has fields for an ID number, number of credit hours earned, and
number of points earned. (For example, many schools compute grade point averages based on a scale of 4, so a
three-credit-hour class in which a student earns an A is worth 12 points.) Include methods to assign values to all
fields. A Student also has a field for grade point average. Include a method to compute the grade point average
field by dividing points by credit hours earned. Write methods to display the values in each Student field. Save
this class as Student.java.

ShowStudent
b. Write a class named ShowStudent that instantiates a Student object from the class you created and assign
values to its fields. Compute the Student grade point average, and then display all the values associated with
the Student. Save the application as ShowStudent.java.

ShowStudent2
c. Create a constructor for the Student class you created. The constructor should initialize each Student’s ID number
to 9999, his or her points earned to 12, and credit hours to 3 (resulting in a grade point average of 4.0). Write a
program that demonstrates that the constructor works by instantiating an object and displaying the initial
values. Save the application as ShowStudent2.java.